requirements_local.txt
# installed Flask-3.0.0 Jinja2-3.1.2 MarkupSafe-2.1.3 Werkzeug-3.0.1 blinker-1.7.0 click-8.1.7 colorama-0.4.6 itsdangerous-2.1.2
Flask==3.0.0
slack_sdk==3.26.0
# installed boto3-1.29.7 botocore-1.32.7 jmespath-1.0.1 python-dateutil-2.8.2 s3transfer-0.8.0 six-1.16.0 urllib3-2.0.7
boto3==1.29.7
# installed certifi-2023.11.17 charset-normalizer-3.3.2 idna-3.6 requests-2.31.0
requests==2.31.0
# installed load_dotenv-0.1.0 python-dotenv-1.0.0
load_dotenv==0.1.0
psycopg2==2.9.9

